Refactor Dockerfile and add docker-compose configuration; update chat-utils for shared access
